+++ src/index.html
@@ -131,7 +131,7 @@
 
 <script src="js/resizer.js" charset="utf-8"></script>
 <script src="js/upload.js" charset="utf-8"></script>
-<script src="js/pictures.js" charset="utf-8"></script>
+<script src="js/main.js" charset="utf-8"></script>
 
 <!--<script src="js/check.js" charset="utf-8"></script>
 <script src="js/upload-stat.min.js" charset="utf-8"></script>-->
