+++ src/js/load.js
@@ -1,0 +1,14 @@
+'use strict';
+
+var load = function(url, callback) {
+  window.callbackName = function(data) {
+    callback(data);
+  };
+
+  var script = document.createElement('script');
+  script.src = url;
+  document.body.appendChild(script);
+
+};
+
+module.exports = load;
