+++ src/js/main.js
@@ -1,1 +1,6 @@
 'use strict';
+
+require('./load');
+require('./picture');
+require('./pictures');
+
