+++ src/js/pictures.js
@@ -1,54 +1,19 @@
 'use strict';
-
-var PICTURE_LOAD_URL = 'http://localhost:1507/api/pictures?callback=callbackName';
 var filters = document.querySelector('.filters');
+var load = require('./load');
+var getPicturesElement = require('./picture');
+var PICTURE_LOAD_URL = 'http://localhost:1507/api/pictures?callback=callbackName';
 
-//var pictures = [];
+(module.exports = function() {
+  var container = document.querySelector('.pictures');
+  filters.classList.add('hidden');
 
-var container = document.querySelector('.pictures');
-var template = document.querySelector('template');
-var templateContainer = 'content' in template ? template.content : template;
-filters.classList.add('hidden');
-
-var getPicturesElement = function(picture) {
-
-  var pictureElement = templateContainer.querySelector('.picture').cloneNode(true);
-  var imgEl = pictureElement.querySelector('img');
-  pictureElement.querySelector('.picture-comments').textContent = picture.comments;
-  pictureElement.querySelector('.picture-likes').textContent = picture.likes;
-
-  var image = new Image(182, 182);
-
-  image.onload = function() {
-    pictureElement.replaceChild(image, imgEl);
+  var renderPictures = function(_pictures) {
+    _pictures.forEach(function(picture) {
+      container.appendChild(getPicturesElement(picture));
+    });
+    filters.classList.remove('hidden');
   };
 
-  image.onerror = function() {
-    pictureElement.classList.add('picture-load-failure');
-  };
-  image.src = picture.preview ? picture.preview : picture.url;
-  return pictureElement;
-};
-
-var renderPictures = function(_pictures) {
-  _pictures.forEach(function(picture) {
-    container.appendChild(getPicturesElement(picture));
-  });
-  filters.classList.remove('hidden');
-};
-
-/*renderPictures(pictures);*/
-
-var load = function(url, callback) {
-  window.callbackName = function(data) {
-    callback(data);
-  };
-
-  var script = document.createElement('script');
-  script.src = url;
-  document.body.appendChild(script);
-
-};
-
-load(PICTURE_LOAD_URL, renderPictures);
-
+  load(PICTURE_LOAD_URL, renderPictures);
+})();
